Refactoring MongoPlugin junits into smaller files
